requirements.txt
fastapi~=0.110.0
pymongo~=4.6.2
requests~=2.31.0
beautifulsoup4~=4.12.3
uvicorn~=0.28.0
python-dotenv~=1.0.1
motor~=3.4.0
aiohttp~=3.9.3
redis~=4.6.0
aioredis~=2.0.1